# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Stargate API
  description: |
    Stargate is a high-performance API gateway that provides comprehensive route management, 
    load balancing, plugin system, and developer portal functionality.
    
    ## Authentication
    
    Stargate supports multiple authentication methods:
    - **API Key**: Use `X-Admin-Key` header for admin operations
    - **JWT Bearer**: Use `Authorization: Bearer <token>` for user operations
    
    ## Rate Limiting
    
    API requests are subject to rate limiting:
    - Admin API: 1000 requests/hour per API key
    - Portal API: 100 requests/minute per user
    
    ## Error Handling
    
    All endpoints return consistent error responses with HTTP status codes and JSON error objects.
  version: 1.0.0
  contact:
    name: Stargate Team
    url: https://github.com/stargate/stargate
    email: support@stargate.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:9090
    description: Development server (Admin API)
  - url: http://localhost:8080
    description: Development server (Portal API)
  - url: https://api.stargate.io
    description: Production server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      summary: Authenticate user
      description: Authenticate user credentials and return JWT token
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              admin:
                summary: Admin login
                value:
                  username: admin
                  password: password
              developer:
                summary: Developer login
                value:
                  username: developer
                  password: dev_password
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/v1/routes:
    get:
      summary: List routes
      description: Get a paginated list of all routes
      operationId: listRoutes
      parameters:
        - name: limit
          in: query
          description: Maximum number of routes to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          description: Number of routes to skip
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: search
          in: query
          description: Search routes by name or host
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of routes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteListResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create route
      description: Create a new route configuration
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
            examples:
              basic:
                summary: Basic route
                value:
                  name: API Route
                  rules:
                    hosts:
                      - api.example.com
                    paths:
                      - type: prefix
                        value: /api
                    methods:
                      - GET
                      - POST
                  upstream_id: upstream-001
              advanced:
                summary: Advanced route with regex
                value:
                  name: User API Route
                  rules:
                    hosts:
                      - api.example.com
                    paths:
                      - type: regex
                        value: ^/api/v1/users/[0-9]+$
                    methods:
                      - GET
                      - PUT
                      - DELETE
                  upstream_id: upstream-001
                  priority: 100
                  metadata:
                    team: backend
                    environment: production
      responses:
        '201':
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict - route already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Admin API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          format: password
          example: password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: string
          description: Token expiration duration
          example: 24h
        refresh_token:
          type: string
          description: Refresh token for obtaining new access tokens
          example: rt_1234567890abcdef
...
    Route:
      type: object
      required:
        - id
        - name
        - rules
        - upstream_id
      properties:
        id:
          type: string
          description: Unique identifier for the route
          example: route-001
        name:
          type: string
          description: Human-readable name for the route
          example: API Route
        rules:
          $ref: '#/components/schemas/RouteRules'
        upstream_id:
          type: string
          description: ID of the upstream service
          example: upstream-001
        priority:
          type: integer
          description: Route priority (higher values have higher priority)
          minimum: 0
          maximum: 1000
          default: 100
          example: 100
        metadata:
          type: object
          description: Additional metadata for the route
          additionalProperties:
            type: string
          example:
            environment: production
            team: backend
        created_at:
          type: integer
          format: int64
          description: Unix timestamp of creation
          example: 1640995200
        updated_at:
          type: integer
          format: int64
          description: Unix timestamp of last update
          example: 1640995200
    RouteRules:
      type: object
      properties:
        hosts:
          type: array
          items:
            type: string
          description: List of host patterns to match
          example:
            - api.example.com
            - "*.api.example.com"
        paths:
          type: array
          items:
            $ref: '#/components/schemas/PathRule'
          description: List of path matching rules
        methods:
          type: array
          items:
            type: string
            enum: [GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS]
          description: List of HTTP methods to match
          example:
            - GET
            - POST
    PathRule:
      type: object
      required:
        - type
        - value
      properties:
        type:
          type: string
          enum: [exact, prefix, regex]
          description: Type of path matching
          example: prefix
        value:
          type: string
          description: Path pattern to match
          example: /api
    RouteRequest:
      type: object
      required:
        - name
        - rules
        - upstream_id
      properties:
        name:
          type: string
          description: Human-readable name for the route
          example: API Route
        rules:
          $ref: '#/components/schemas/RouteRules'
        upstream_id:
          type: string
          description: ID of the upstream service
          example: upstream-001
        priority:
          type: integer
          description: Route priority
          minimum: 0
          maximum: 1000
          default: 100
          example: 100
        metadata:
          type: object
          description: Additional metadata
          additionalProperties:
            type: string
    RouteResponse:
      type: object
      properties:
        message:
          type: string
          example: Route created successfully
        route:
          $ref: '#/components/schemas/Route'
    RouteListResponse:
      type: object
      properties:
        routes:
          type: array
          items:
            $ref: '#/components/schemas/Route'
        total:
          type: integer
          description: Total number of routes
          example: 150
        limit:
          type: integer
          description: Maximum number of routes returned
          example: 50
        offset:
          type: integer
          description: Number of routes skipped
          example: 0
        has_more:
          type: boolean
          description: Whether there are more routes available
          example: true
...
    Error:
      type: object
      required:
        - error
        - status
      properties:
        error:
          type: string
          description: Error message
          example: Validation failed
        status:
          type: integer
          description: HTTP status code
          example: 400
        details:
          type: string
          description: Additional error details
          example: Field 'name' is required
        code:
          type: string
          description: Error code for programmatic handling
          example: VALIDATION_ERROR
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp of the error
          example: 1640995200
...